requirements.txt
asgiref==3.9.1
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
cryptography==45.0.7
dj-database-url==3.0.1
Django==5.2.6
django-allauth==65.11.2
django-environ==0.12.0
djangorestframework==3.16.1
drf-yasg==1.21.10
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.10
pycparser==2.23
PyJWT==2.10.1
python-decouple==3.8
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
reportlab==4.2.5
requests==2.32.5
sqlparse==0.5.3
uritemplate==4.2.0
urllib3==2.5.0
whitenoise==6.10.0
